move base_mesh to base_regular_mesh
