+++ discretize/base/__init__.py
@@ -16,5 +16,5 @@ Base Mesh Class
   BaseTensorMesh
 """
 
-from discretize.base.base_mesh import BaseMesh, BaseRectangularMesh
+from discretize.base.base_regular_mesh import BaseMesh, BaseRectangularMesh
 from discretize.base.base_tensor_mesh import BaseTensorMesh
